+++ moveit_planners/ompl/ompl_interface/src/ompl_planner_manager.cpp
@@ -150,114 +150,6 @@ public:
   }
 
 private:
-  /*
-  bool r = ompl_interface_->solve(planning_scene, req, res);
-  if (!planner_data_link_name_.empty())
-    displayPlannerData(planning_scene, planner_data_link_name_);
-  return r;
-  */
-
-  /*
-  bool r = ompl_interface_->solve(planning_scene, req, res);
-  if (!planner_data_link_name_.empty())
-    displayPlannerData(planning_scene, planner_data_link_name_);
-  return r;
-  */
-
-  /*
-  void displayRandomValidStates()
-  {
-    ompl_interface::ModelBasedPlanningContextPtr pc = ompl_interface_->getLastPlanningContext();
-    if (!pc || !pc->getPlanningScene())
-    {
-      ROS_ERROR("No planning context to sample states for");
-      return;
-    }
-    ROS_INFO_STREAM("Displaying states for context " << pc->getName());
-    const og::SimpleSetup &ss = pc->getOMPLSimpleSetup();
-    ob::ValidStateSamplerPtr vss = ss.getSpaceInformation()->allocValidStateSampler();
-    robot_state::RobotState kstate = pc->getPlanningScene()->getCurrentState();
-    ob::ScopedState<> rstate1(ss.getStateSpace());
-    ob::ScopedState<> rstate2(ss.getStateSpace());
-    ros::WallDuration wait(2);
-    unsigned int n = 0;
-    std::vector<ob::State*> sts;
-    if (vss->sample(rstate2.get()))
-      while (display_random_valid_states_)
-      {
-        if (!vss->sampleNear(rstate1.get(), rstate2.get(), 10000000))
-          continue;
-        pc->getOMPLStateSpace()->copyToRobotState(kstate, rstate1.get());
-        kstate.getJointStateGroup(pc->getJointModelGroupName())->updateLinkTransforms();
-        moveit_msgs::DisplayRobotState state_msg;
-        robot_state::robotStateToRobotStateMsg(kstate, state_msg.state);
-        pub_valid_states_.publish(state_msg);
-        n = (n + 1) % 2;
-        if (n == 0)
-        {
-          robot_trajectory::RobotTrajectory traj(pc->getRobotModel(), pc->getJointModelGroupName());
-          unsigned int g = ss.getSpaceInformation()->getMotionStates(rstate1.get(), rstate2.get(), sts, 10, true, true);
-          ROS_INFO("Generated a motion with %u states", g);
-          for (std::size_t i = 0 ; i < g ; ++i)
-          {
-            pc->getOMPLStateSpace()->copyToRobotState(kstate, sts[i]);
-            traj.addSuffixWayPoint(kstate, 0.0);
-          }
-          moveit_msgs::DisplayTrajectory msg;
-          msg.model_id = pc->getRobotModel()->getName();
-          msg.trajectory.resize(1);
-          traj.getRobotTrajectoryMsg(msg.trajectory[0]);
-          robot_state::robotStateToRobotStateMsg(traj.getFirstWayPoint(), msg.trajectory_start);
-          pub_valid_traj_.publish(msg);
-        }
-        rstate2 = rstate1;
-        wait.sleep();
-      }
-  }
-
-  void displayPlannerData(const planning_scene::PlanningSceneConstPtr& planning_scene,
-                          const std::string &link_name) const
-  {
-    ompl_interface::ModelBasedPlanningContextPtr pc = ompl_interface_->getLastPlanningContext();
-    if (pc)
-    {
-      ompl::base::PlannerData pd(pc->getOMPLSimpleSetup()->getSpaceInformation());
-      pc->getOMPLSimpleSetup()->getPlannerData(pd);
-      robot_state::RobotState kstate = planning_scene->getCurrentState();
-      visualization_msgs::MarkerArray arr;
-      std_msgs::ColorRGBA color;
-      color.r = 1.0f;
-      color.g = 0.25f;
-      color.b = 1.0f;
-      color.a = 1.0f;
-      unsigned int nv = pd.numVertices();
-      for (unsigned int i = 0 ; i < nv ; ++i)
-      {
-        pc->getOMPLStateSpace()->copyToRobotState(kstate, pd.getVertex(i).getState());
-        kstate.getJointStateGroup(pc->getJointModelGroupName())->updateLinkTransforms();
-        const Eigen::Vector3d &pos = kstate.getLinkState(link_name)->getGlobalLinkTransform().translation();
-
-        visualization_msgs::Marker mk;
-        mk.header.stamp = ros::Time::now();
-        mk.header.frame_id = planning_scene->getPlanningFrame();
-        mk.ns = "planner_data";
-        mk.id = i;
-        mk.type = visualization_msgs::Marker::SPHERE;
-        mk.action = visualization_msgs::Marker::ADD;
-        mk.pose.position.x = pos.x();
-        mk.pose.position.y = pos.y();
-        mk.pose.position.z = pos.z();
-        mk.pose.orientation.w = 1.0;
-        mk.scale.x = mk.scale.y = mk.scale.z = 0.025;
-        mk.color = color;
-        mk.lifetime = ros::Duration(30.0);
-        arr.markers.push_back(mk);
-      }
-      pub_markers_.publish(arr);
-    }
-  }
-  */
-
   void dynamicReconfigureCallback(OMPLDynamicReconfigureConfig& config, uint32_t level)
   {
     if (config.link_for_exploration_tree.empty() && !planner_data_link_name_.empty())
